Added LoggingIn and CreateAccount functionality
Added Logging in for an existing user and also added the create account
for new users. You can create the user as any of the roles in the
database. New users i added are kate.anglin@email.com and password is
the password and organizer@email.com and password is password. The
passwords are encrypted and safe. Had to delete/comment out lots of code
that was from previous project manager implementation to make work.