Created the OrganizerPortfolio and Create Event
I created the UI for the event and the Organizers Profile. This now allows the organizer to log in and see a unique page for the organizer only. This allows the organizer to create new event and view thier event. I made it where random event appear on the home screen and am working on the view for anyone to read about the event details. All I have left in make the dates work and implmenting a way to geth the users name with only the ID.